+++ src/main/java/com/toyboard/jeongmin/post/domain/BaseTimeEntity.java
@@ -1,4 +1,4 @@
-package com.toyboard.jeongmin.domain;
+package com.toyboard.jeongmin.post.domain;
 
 import lombok.Getter;
 import org.springframework.data.annotation.CreatedDate;
