+++ src/main/java/com/toyboard/jeongmin/post/request/PostRequest.java
@@ -1,4 +1,4 @@
-package com.toyboard.jeongmin.request;
+package com.toyboard.jeongmin.post.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
